apps/api/agent-worker/requirements.txt
livekit-agents[openai]~=1.2
python-dotenv>=1.0
